Fix type

+++ src/common/common/parser.cpp
@@ -58,14 +58,14 @@ void split_to_ngrams(
       list_strings.end(),
       [&ngram_words, &ngram_min_length, &ngram_max_length, &ngram_index](
           std::string &item) {
-        const uint64_t size = item.size();
+        const size_t size = item.size();
         if (size < ngram_max_length && size > ngram_min_length) {
-          for (uint64_t i = ngram_min_length; i <= size; i++) {
+          for (size_t i = ngram_min_length; i <= size; i++) {
             ngram_words.insert({item.substr(0, i), ngram_index});
           }
           ngram_index++;
         } else if (size >= ngram_max_length) {
-          for (uint16_t i = ngram_min_length; i <= ngram_max_length; i++) {
+          for (size_t i = ngram_min_length; i <= ngram_max_length; i++) {
             ngram_words.insert({item.substr(0, i), ngram_index});
           }
           ngram_index++;
